+++ rust/main/agents/relayer/src/msg/gas_payment/policies/on_chain_fee_quoting.rs
@@ -41,7 +41,6 @@ impl GasPaymentPolicy for GasPaymentPolicyOnChainFeeQuoting {
     /// OnChainFeeQuoting requires the user to pay a specified fraction of the
     /// estimated gas. Like the Minimum policy, OnChainFeeQuoting requires a
     /// payment to exist on the IGP specified in the config.
-
     async fn message_meets_gas_payment_requirement(
         &self,
         _message: &HyperlaneMessage,
@@ -49,9 +48,11 @@ impl GasPaymentPolicy for GasPaymentPolicyOnChainFeeQuoting {
         current_expenditure: &InterchainGasExpenditure,
         tx_cost_estimate: &TxCostEstimate,
     ) -> Result<Option<U256>> {
-        let fractional_gas_estimate = (tx_cost_estimate.enforceable_gas_limit()
-            * self.fractional_numerator)
-            / self.fractional_denominator;
+        let fractional_gas_estimate = tx_cost_estimate
+            .enforceable_gas_limit()
+            .saturating_mul(U256::from(self.fractional_numerator))
+            .div_mod(U256::from(self.fractional_denominator))
+            .0;
         let gas_amount = current_payment
             .gas_amount
             .saturating_sub(current_expenditure.gas_used);