+++ typescript/sdk/src/token/Token.ts
@@ -18,6 +18,7 @@ import type { IToken, TokenArgs } from './IToken.js';
 import { TokenAmount } from './TokenAmount.js';
 import { TokenConnection, TokenConnectionType } from './TokenConnection.js';
 import {
+  PROTOCOL_TO_HYP_NATIVE_STANDARD,
   PROTOCOL_TO_NATIVE_STANDARD,
   TOKEN_COLLATERALIZED_STANDARDS,
   TOKEN_HYP_STANDARDS,
@@ -59,6 +60,11 @@ import type {
   IHypTokenAdapter,
   ITokenAdapter,
 } from './adapters/ITokenAdapter.js';
+import {
+  RadixHypCollateralAdapter,
+  RadixHypSyntheticAdapter,
+  RadixNativeTokenAdapter,
+} from './adapters/RadixTokenAdapter.js';
 import {
   SealevelHypCollateralAdapter,
   SealevelHypNativeAdapter,
@@ -156,6 +162,10 @@ export class Token implements IToken {
         {},
         addressOrDenom,
       );
+    } else if (standard === TokenStandard.RadixNative) {
+      return new RadixNativeTokenAdapter(chainName, multiProvider, {
+        token: addressOrDenom,
+      });
     } else if (this.isHypToken()) {
       return this.getHypAdapter(multiProvider);
     } else if (this.isIbcToken()) {
@@ -319,6 +329,14 @@ export class Token implements IToken {
       return new StarknetHypCollateralAdapter(chainName, multiProvider, {
         warpRouter: addressOrDenom,
       });
+    } else if (standard === TokenStandard.RadixHypCollateral) {
+      return new RadixHypCollateralAdapter(chainName, multiProvider, {
+        token: addressOrDenom,
+      });
+    } else if (standard === TokenStandard.RadixHypSynthetic) {
+      return new RadixHypSyntheticAdapter(chainName, multiProvider, {
+        token: addressOrDenom,
+      });
     } else {
       throw new Error(`No hyp adapter found for token standard: ${standard}`);
     }
@@ -387,6 +405,12 @@ export class Token implements IToken {
 
   isNative(): boolean {
     return Object.values(PROTOCOL_TO_NATIVE_STANDARD).includes(this.standard);
+  }
+
+  isHypNative(): boolean {
+    return Object.values(PROTOCOL_TO_HYP_NATIVE_STANDARD).includes(
+      this.standard,
+    );
   }
 
   isCollateralized(): boolean {